refactor "recover" domain, improve README
